requirements.txt
cramjam==2.11.0
fastparquet==2024.11.0
filelock==3.20.0
fsspec==2025.10.0
huggingface-hub>=0.34,<1.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pyarrow==22.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.2
hmmlearn==0.3.3
chronos-forecasting
git+https://github.com/shiyu-coder/Kronos.git
